labels2edges bug fix.

+++ functions/labels2edges.m
@@ -2,10 +2,13 @@
 % Converts a labeled segmentation to segmentation boundaries
 function edges = labels2edges(labels)
 
-grad = imfilter(labels,[-1 3; -1 -1]);
+grad = imfilter(double(labels),[-1 3; -1 -1]);
 edges = abs(grad)>0;
 
 % edges = edge(labels,0);
+
+% [GX,GY] = gradient(double(labels));
+% edges = logical(min((abs(GX)>0)+(abs(GY)>0),1));
 
 end
